some changes and package distribution
