--- src/Mode.java
@@ -1,4 +1,0 @@
-enum Mode {
-    CODING,
-    DECODING
-}
