+++ src/parameters/Mode.java
@@ -1,0 +1,6 @@
+package parameters;
+
+public enum Mode {
+    CODING,
+    DECODING
+}
