+++ src/parameters/Type.java
@@ -1,4 +1,6 @@
-enum Type {
+package parameters;
+
+public enum Type {
     STRING,
     STRINGBUILDER,
     COMPARABLE
